Refactor makeResponsive and utils/css
- Split utils.css out of utils/react
- Allow makeResponsive to accept arbitrary media query
