+++ v3/src/js/utils/react.jsx
@@ -2,43 +2,12 @@
 
 import type { Node, ComponentType } from 'react';
 import React from 'react';
-import { escapeRegExp, castArray, entries } from 'lodash';
+import { escapeRegExp, castArray } from 'lodash';
 
 // Define some useful Unicode characters as constants
 export const NBSP = '\u00a0';
 export const BULLET = ' • ';
 
-// Define media breakpoints
-export type Breakpoint = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
-const breakpoints: { [Breakpoint]: number } = {
-  xs: 0,
-  sm: 576,
-  md: 768,
-  lg: 992,
-  xl: 1200,
-};
-
-function nextBreakpoint(size: Breakpoint): ?Breakpoint {
-  const breakpointEntries = entries(breakpoints);
-  const nextBreakpointIndex = breakpointEntries.findIndex(([breakpoint]) => breakpoint === size) + 1;
-  if (nextBreakpointIndex >= breakpointEntries.length) return null;
-  return breakpointEntries[nextBreakpointIndex][1];
-}
-
-export function breakpointDown(size: Breakpoint): MediaQueryList {
-  const nextSize = nextBreakpoint(size);
-  if (nextSize == null) return window.matchMedia('all');
-  return window.matchMedia(`(max-width: ${breakpoints[nextSize] - 1}px )`);
-}
-
-export function breakpointUp(size: Breakpoint): MediaQueryList {
-  return window.matchMedia(`(min-width: ${breakpoints[size]}px)`);
-}
-
-export function supportsCSSVariables() {
-  // Safari does not support supports('--var', 'red')
-  return window.CSS.supports && window.CSS.supports('(--var: red)');
-}
 
 /**
  * Replace substring matching the provided regex with React nodes. This is
